changing variable type

+++ GEOSagcm_GridComp/GEOSphysics_GridComp/GEOSsurface_GridComp/GEOSland_GridComp/GEOScatchCN_GridComp/GEOScatchCNCLM51_GridComp/CLM51/CNCLM_DriverMod.F90
@@ -663,7 +663,7 @@ contains
   ! LOCAL
   integer :: n, p, nv, nc, nz, np
 
-  real, pointer :: elai_clm(:)
+  real(r8), pointer :: elai_clm(:)
   !------------------------------
 
   elai_clm => canopystate_inst%elai_patch
